addDataFrame.Rd documentation RE:date/time format
I struggled to interpret the documentation for the addDataFrame function in relation to date time formats when writing xlsx.

Outside of Rmarkdown I rarely see the options() function used to control the output of functions. So it wasn't immediately clear that this was what I was supposed to use to control date formats within xlsx.

I have made a suggestion as to how this might be more explicitly communicated.